requirements.txt
astroid==2.3.3
autopep8==1.5
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
lxml==4.5.0
mccabe==0.6.1
psycopg2-binary==2.8.4
pycodestyle==2.5.0
pylint==2.4.4
python-dateutil==2.8.1
pytz==2019.3
regex==2020.2.20
requests==2.23.0
rope==0.16.0
six==1.14.0
soupsieve==1.9.5
tzlocal==2.0.0
Unidecode==1.1.1
urllib3==1.25.8
wrapt==1.11.2
